refactor(test): factorize test app
